+++ askbot/mail/lamson_handlers.py
@@ -7,7 +7,7 @@ from django.utils.translation import ugettext as _
 from lamson.routing import route, stateless
 from lamson.server import Relay
 from askbot.models import ReplyAddress, Tag
-from askbot.utils import mail
+from askbot import mail
 from askbot.conf import settings as askbot_settings
 from askbot.skins.loaders import get_template
 
@@ -187,9 +187,13 @@ def ASK(message, host = None, addr = None):
 
     parts = get_parts(message)
     from_address = message.From
-    subject = message['Subject']#why lamson does not give it normally?
+    #why lamson does not give it normally?
+    subject = message['Subject'].strip('\n\t ')
+    body_text, stored_files, unused = mail.process_parts(parts)
     if addr == 'ask':
-        mail.process_emailed_question(from_address, subject, parts)
+        mail.process_emailed_question(
+            from_address, subject, body_text, stored_files
+        )
     else:
         if askbot_settings.GROUP_EMAIL_ADDRESSES_ENABLED == False:
             return
@@ -199,7 +203,8 @@ def ASK(message, host = None, addr = None):
                 name__iexact = addr
             )
             mail.process_emailed_question(
-                from_address, subject, parts, tags = [group_tag.name, ]
+                from_address, subject, body_text, stored_files,
+                tags = [group_tag.name, ]
             )
         except Tag.DoesNotExist:
             #do nothing because this handler will match all emails
@@ -220,28 +225,12 @@ def VALIDATE_EMAIL(
     the email signature
     todo: go a step further and
     """
-    content, stored_files = mail.process_parts(parts)
     reply_code = reply_address_object.address
-    if reply_code in content:
-
-        #extract the signature
-        tail = list()
-        for line in reversed(content.splitlines()):
-            #scan backwards from the end until the magic line
-            if reply_code in line:
-                break
-            tail.insert(0, line)
-
-        #strip off the leading quoted lines, there could be one or two
-        #also strip empty lines
-        while tail[0].startswith('>') or tail[0].strip() == '':
-            tail.pop(0)
-
-        signature = '\n'.join(tail)
-
-        #save the signature and mark email as valid
+    try:
+        content, stored_files, signature = mail.process_parts(parts, reply_code)
         user = reply_address_object.user
-        user.email_signature = signature
+        if signature and signature != user.email_signature:
+            user.email_signature = signature
         user.email_isvalid = True
         user.save()
 
@@ -257,8 +246,7 @@ def VALIDATE_EMAIL(
             body_text = template.render(Context(data)),
             recipient_list = [from_address,]
         )
-
-    else:
+    except ValueError:
         raise ValueError(
             _(
                 'Please reply to the welcome email '
@@ -272,12 +260,58 @@ def VALIDATE_EMAIL(
 def PROCESS(
     parts = None,
     reply_address_object = None,
+    subject_line = None,
+    from_address = None,
     **kwargs
 ):
     """handler to process the emailed message
     and make a post to askbot based on the contents of
     the email, including the text body and the file attachments"""
-    if reply_address_object.was_used:
-        reply_address_object.edit_post(parts)
-    else:
-        reply_address_object.create_reply(parts)
+    #1) get actual email content 
+    #   todo: factor this out into the process_reply decorator
+    reply_code = reply_address_object.address
+    body_text, stored_files, signature = mail.process_parts(parts, reply_code)
+
+    #2) process body text and email signature
+    user = reply_address_object.user
+    if signature:#if there, then it was stripped
+        if signature != user.email_signature:
+            user.email_signature = signature
+    else:#try to strip signature
+        stripped_body_text = user.strip_email_signature(body_text)
+        #todo: add test cases for emails without the signature
+        if stripped_body_text == body_text and user.email_signature:
+            #todo: send an email asking to update the signature
+            raise ValueError('email signature changed or unknown')
+        body_text = stripped_body_text
+
+    #3) validate email address and save user
+    user.email_isvalid = True
+    user.save()#todo: actually, saving is not necessary, if nothing changed
+
+    #4) actually make an edit in the forum
+    robj = reply_address_object
+    add_post_actions = ('post_comment', 'post_answer', 'auto_answer_or_comment')
+    if robj.reply_action == 'replace_content':
+        robj.edit_post(body_text, title = subject_line)
+    elif robj.reply_action == 'append_content':
+        robj.edit_post(body_text)#in this case we don't touch the title
+    elif robj.reply_action in add_post_actions:
+        if robj.was_used:
+            robj.edit_post(body_text, edit_response = True)
+        else:
+            robj.create_reply(body_text)
+    elif robj.reply_action == 'validate_email':
+        #todo: this is copy-paste - factor it out to askbot.mail.messages
+        data = {
+            'site_name': askbot_settings.APP_SHORT_NAME,
+            'site_url': askbot_settings.APP_URL,
+            'ask_address': 'ask@' + askbot_settings.REPLY_BY_EMAIL_HOSTNAME
+        }
+        template = get_template('email/re_welcome_lamson_on.html')
+
+        mail.send_mail(
+            subject_line = _('Re: %s') % subject_line,
+            body_text = template.render(Context(data)),
+            recipient_list = [from_address,]
+        )