+++ askbot/models/reply_by_email.py
@@ -4,11 +4,11 @@ import string
 import logging
 from django.db import models
 from django.contrib.auth.models import User
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import ugettext as _
 from askbot.models.post import Post
 from askbot.models.base import BaseQuerySetManager
 from askbot.conf import settings as askbot_settings
-from askbot.utils import mail
+from askbot import mail
 
 class ReplyAddressManager(BaseQuerySetManager):
     """A manager for the :class:`ReplyAddress` model"""
@@ -36,6 +36,8 @@ class ReplyAddressManager(BaseQuerySetManager):
 REPLY_ACTION_CHOICES = (
     ('post_answer', _('Post an answer')),
     ('post_comment', _('Post a comment')),
+    ('replace_content', _('Edit post')),
+    ('append_content', _('Append to post')),
     ('auto_answer_or_comment', _('Answer or comment, depending on the size of post')),
     ('validate_email', _('Validate email and record signature')),
 )
@@ -74,36 +76,69 @@ class ReplyAddress(models.Model):
         """True if was used"""
         return self.used_at != None
 
-    def edit_post(self, parts):
+    def as_email_address(self, prefix = 'reply-'):
+        """returns email address, prefix is added
+        in front of the code"""
+        return '%s%s@%s' % (
+                        prefix,
+                        self.address,
+                        askbot_settings.REPLY_BY_EMAIL_HOSTNAME
+                    )
+
+    def edit_post(
+        self, body_text, title = None, edit_response = False
+    ):
         """edits the created post upon repeated response
         to the same address"""
-        assert self.was_used == True
-        content, stored_files = mail.process_parts(parts)
-        self.user.edit_post(
-            post = self.response_post,
-            body_text = content,
-            revision_comment = _('edited by email'),
-            by_email = True
-        )
-        self.response_post.thread.invalidate_cached_data()
+        if self.was_used or edit_response:
+            reply_action = 'append_content'
+        else:
+            reply_action = self.reply_action
 
-    def create_reply(self, parts):
+        if edit_response:
+            post = self.response_post
+        else:
+            post = self.post
+
+        if reply_action == 'append_content':
+            body_text = post.text + '\n\n' + body_text
+            revision_comment = _('added content by email')
+        else:
+            assert(reply_action == 'replace_content')
+            revision_comment = _('edited by email')
+
+        if post.post_type == 'question':
+            assert(post is self.post)
+            self.user.edit_question(
+                question = post,
+                body_text = body_text,
+                title = title,
+                revision_comment = revision_comment,
+                by_email = True
+            )
+        else:
+            self.user.edit_post(
+                post = post,
+                body_text = body_text,
+                revision_comment = revision_comment,
+                by_email = True
+            )
+        self.post.thread.invalidate_cached_data()
+
+    def create_reply(self, body_text):
         """creates a reply to the post which was emailed
         to the user
         """
         result = None
-        #todo: delete stored files if this function fails
-        content, stored_files = mail.process_parts(parts)
-
         if self.post.post_type == 'answer':
             result = self.user.post_comment(
                                         self.post,
-                                        content,
+                                        body_text,
                                         by_email = True
                                     )
         elif self.post.post_type == 'question':
             if self.reply_action == 'auto_answer_or_comment':
-                wordcount = len(content)/6#todo: this is a simplistic hack
+                wordcount = len(body_text)/6#todo: this is a simplistic hack
                 if wordcount > askbot_settings.MIN_WORDS_FOR_ANSWER_BY_EMAIL:
                     reply_action = 'post_answer'
                 else:
@@ -114,13 +149,13 @@ class ReplyAddress(models.Model):
             if reply_action == 'post_answer':
                 result = self.user.post_answer(
                                             self.post,
-                                            content,
+                                            body_text,
                                             by_email = True
                                         )
             elif reply_action == 'post_comment':
                 result = self.user.post_comment(
                                             self.post,
-                                            content,
+                                            body_text,
                                             by_email = True
                                         )
             else:
@@ -131,7 +166,7 @@ class ReplyAddress(models.Model):
         elif self.post.post_type == 'comment':
             result = self.user.post_comment(
                                     self.post.parent,
-                                    content,
+                                    body_text,
                                     by_email = True
                                 )
         result.thread.invalidate_cached_data()